+++ packages/shared/src/libs/helpers/helpers.ts
@@ -4,5 +4,6 @@ export { checkHasPermission } from "./check-has-permission/check-has-permission.
 export { configureString } from "./configure-string/configure-string.helper.js";
 export { formatDate } from "./format-date/format-date.helper.js";
 export { getDifferenceInDays } from "./get-difference-in-days/get-difference-in-days.helper.js";
+export { getRelativeDate } from "./get-relative-date/get-relative-date.helper.js";
 export { initDebounce } from "./init-debounce/init-debounce.helper.js";
 export { subtractDays } from "./subtract-days/subtract-days.helper.js";