+++ packages/shared/src/modules/project-groups/libs/types/project-group-patch-request-dto.type.ts
@@ -1,0 +1,7 @@
+type ProjectGroupPatchRequestDto = {
+	name: string;
+	permissionIds: number[];
+	userIds: number[];
+};
+
+export { type ProjectGroupPatchRequestDto };